shithead: fix player copy.
In MCTS each node needs its own copy of the game state.
Unfortunately I accomplished this by making a copy of the game state
inside Game's next_state() methode. This has broken card swapping for AI
players, because I've forgotten to include the swap_count attribut in
the copy() method of the AI players (i.e. swap_count was reset to 0 when
next_state() created a new copy of the state and thus of the players.
Finding the cards from the best_fup list in the AI player's hand did
also not work with copies (compared pointer instead of card attributes).
I fixed all this problems, so AI card swapping would now also work with
state copies, but also reversed the change to next_state() and added
copying the state to the MonteCarlo expand() method.