shithead: replace DeepShit with DeeperShit.
DeeperShit is an improved version of DeepShit using MTCS for the end
game, but instead of providing a new AI type in the GUI to the user, we
just silently replace DeepShit with DeeperShit (except of the MTCS part,
they are both quite similar). I.e. the user selects DeepShit, but behind
the scene, DeeperShit selects the plays.
With the use of DeeperShit an existing Problem became obvious.
At the end of the game after the shithead has been found and the user
continues to the result screen with a mouse click, another play
selection is requested from the only player left. In case of DeeperShit
an MCTS thread (search for 1 s) is started, which will never be properly
terminated, and which will bring the GUI to a grinding halt.
So I fixed this by allowing get_play() only to be called if at least 2
players are still in the players list.