shithead: fix play marking.
The bright green frames marking legal plays for the human player are
drawn while cards from the play of the previous player are still moving.
This is especially confusing if the previous player has to take the
discard pile. I fixed this by not drawing these frames as long as the
'started' flag of the CardMover is set.
I also blocked drawing frames if less than 2 players are left, to avoid
that legal plays are marked after the human player has become Shithead.