shithead: initialize round result.
The result in the game state keeps for every player the score and the
turn count for the ongoing round. This is used to revert the statistics
in case of a game abort.
During state initialization we create a result dictionary using the
player names as keys an values set to [0,0].