shithead: add reverting of statistics.
During a game round score, turn count, and game count is updated in the
statistics whenever a player goes out. In case of an aborted game round
we have to revert the statistics to the previous values.
There's now also the possibility that we have to print a result table
with 0 games, i.e. we have to avoid division through zero.