shithead: get dealer from statistics.
Because the result may be reverted to the previous round in case of a
game abort, we can no longer use the shithead from the current round as
dealer of the next round. Instead we save the last known shithead in the
statistics and result.py uses this to setup the next round.