shithead: handle game abort.
When a player issues the 'ABORT' play (too many turns => avoid AI
deadlocks), it is now handled with a message and after a mouse click,
the result screen with the statistics reverted to the previous round is
displayed.
I also fixed the bug, where another play is requested from the shithead
(or the next player in an aborted game) by adding a missing return in
the on_update() methode of the game view.